Rename Probably an upland.py to Mislabeled_upland.py
The purpose of this script is to identify layers which are likely upland and have been mislabeled. It does this by using a custom SQL expression to select features by attribute where area of the polygon is greater than 150 acres and the polygon is not labeled as upland. The user is then able to evaluate selected polygons and determine if they are mislabeled.